feat: refactors AI provider interfaces for improved organization
