+++ apps/cli/src/ai-providers/index.ts
@@ -1,2 +1,1 @@
-export * from './ai-provider.interface';
 export * from './ai-provider-registry.service';
