+++ packages/agent/src/ai/ai-provider.interface.ts
@@ -44,6 +44,40 @@ export interface AiProviderCapabilities {
 	};
 }
 
+// CLI-specific interfaces for setup and configuration
+export interface AiProviderDefaults {
+    model: string;
+    temperature: number;
+    maxTokens: number;
+    baseUrl?: string;
+}
+
+export interface AiProviderInfo {
+    name: string;
+    displayName: string;
+    description: string;
+    defaults: AiProviderDefaults;
+    requiresApiKey: boolean;
+    models: string[];
+    websiteUrl: string;
+    docsUrl: string;
+}
+
+export interface ConfiguredAiProvider {
+    name: string;
+    apiKey: string;
+    model: string;
+    temperature: number;
+    maxTokens: number;
+    baseUrl?: string;
+}
+
+export interface AiProviderConfiguration {
+    defaultProvider: string;
+    fallbackProviders: string[];
+    providers: ConfiguredAiProvider[];
+}
+
 export const AI_PROVIDER_CAPABILITIES: Record<AiProviderType, AiProviderCapabilities> = {
 	openai: {
 		supportsStructuredOutput: true,
